Struct governance_os_pallet_tokens::NativeCurrencyAdapter

pub struct NativeCurrencyAdapter<Pallet, GetCurrencyId>(_);

This struct is useful to implement the Currency trait for any given currency inside the system. It basically takes an interface to the tokens pallet and a CurrencyId to expose them under the Currency trait.