Trait governance_os_pallet_tokens::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + Into<Self::Event>;
    type CurrencyId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + Default;
    type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize;
    type WeightInfo: WeightInfo;
    type RoleManager: RoleManager<AccountId = Self::AccountId>;
    type RoleBuilder: RoleBuilder<CurrencyId = Self::CurrencyId, Role = <Self::RoleManager as RoleManager>::Role>;
}

Associated Types

type Event: From<Event<Self>> + Into<Self::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type CurrencyId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + Default

The type used to identify currencies

type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize

The balance of an account.

type WeightInfo: WeightInfo

Weight values for this pallet

type RoleManager: RoleManager<AccountId = Self::AccountId>

Pallet that is in charge of managing the roles based ACL.

type RoleBuilder: RoleBuilder<CurrencyId = Self::CurrencyId, Role = <Self::RoleManager as RoleManager>::Role>

This pallet relies on roles associated to a specific metadata so we need the runtime to provide some helper functions to build those so that we can keep the role definition code modular.

Loading content...