Trait governance_os_pallet_tokens::RoleBuilder

pub trait RoleBuilder {
    type CurrencyId;
    type Role;
    fn transfer_currency(id: Self::CurrencyId) -> Self::Role;

    fn manage_currency(id: Self::CurrencyId) -> Self::Role;

    fn create_currencies() -> Self::Role;
}

Associated Types

type CurrencyId

type Role

Loading content...
Required methods

fn transfer_currency(id: Self::CurrencyId) -> Self::Role

Role for transferring certain units of currencies. This is used to create non transferrable assets or could be used for creating permissioned assets in the future.

fn manage_currency(id: Self::CurrencyId) -> Self::Role

Role for the account(s) that are allowed to mint or burn units of currency.

fn create_currencies() -> Self::Role

Role for creating currencies.

Loading content...